agents/tester/sous_agents/unit_tester/agent.py
"""
Sous-agent Unit Testing
Spécialisation: Unit Testing
"""
from typing import Dict, Any
import yaml
import logging

class UnitTesterSubAgent:
    """Sous-agent spécialisé en Unit Testing"""
    
    def __init__(self, config_path: str = ""):
        self.config_path = config_path
        self.config = self._load_config()
        self.logger = logging.getLogger(__name__)
        self.agent_id = f"unit_tester_sub_01"
        
        self.logger.info(f"Sous-agent {self.agent_id} initialisé")
    
    def _load_config(self) -> Dict[str, Any]:
        """Charge la configuration"""
        if self.config_path and os.path.exists(self.config_path):
            try:
                with open(self.config_path, 'r') as f:
                    return yaml.safe_load(f)
            except Exception as e:
                self.logger.error(f"Erreur de chargement config: {e}")
        
        # Configuration par défaut
        return {
            "agent": {
                "name": "Unit Testing",
                "specialization": "Unit Testing",
                "version": "1.0.0"
            },
            "capabilities": ["task_execution", "specialized_operation"]
        }
    
    async def execute(self, task_data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
        """Exécute une tâche spécialisée"""
        task_type = task_data.get("task_type", "unknown")
        
        self.logger.info(f"Exécution de la tâche {task_type}")
        
        # Implémentation spécifique au sous-agent
        result = await self._execute_specialized(task_data, context)
        
        return {
            "success": True,
            "sub_agent": "unit_tester",
            "task": task_type,
            "result": result,
            "specialization": "Unit Testing"
        }
    
    async def _execute_specialized(self, task_data: Dict[str, Any], context: Dict[str, Any]) -> Dict[str, Any]:
        """Méthode spécialisée à implémenter"""
        # À implémenter selon la spécialisation
        return {
            "message": "Tâche exécutée par le sous-agent spécialisé",
            "specialization": "Unit Testing",
            "input": task_data
        }
    
    async def health_check(self) -> Dict[str, Any]:
        """Vérifie la santé du sous-agent"""
        return {
            "agent": "unit_tester",
            "status": "healthy",
            "type": "sub_agent",
            "specialization": "Unit Testing",
            "config_loaded": bool(self.config)
        }
    
    def get_agent_info(self) -> Dict[str, Any]:
        """Retourne les informations du sous-agent"""
        return {
            "id": self.agent_id,
            "name": "Unit Testing",
            "type": "sub_agent",
            "parent": "tester",
            "specialization": "Unit Testing",
            "version": "1.0.0"
        }
